(SBL-270) Rework error handling

+++ api/src/services/user/routes.js
@@ -1,41 +1,25 @@
 import objection, { raw } from 'objection';
 
-import config from '../../../config';
-
-import validatorCompiler from '../../validation/validatorCompiler';
 import {
   AuthorizationError,
-  BadRequestError,
   NotFoundError,
-  UniqueViolationError,
+  BadRequestError,
+  ConflictError,
 } from '../../validation/errors';
 
-import {
-  signupBodyValidator,
-  signinBodyValidator,
-  patchBodyValidator,
-  validateId,
-  roleBodyValidator,
-  refreshBodyValidator,
-} from './validators';
 import { createAccessToken, createRefreshToken } from './utils';
-import {
-  UNAUTHORIZED,
-  USER_ALREADY_REGISTERED,
-  USER_NOT_FOUND,
-  USER_ADMIN_FIELDS,
-  USER_DELETED,
-  INVALID_PATCH,
-  ALTER_ROLE_SUCCESS,
-  ALTER_ROLE_FAIL,
-  USER_ROLE_NOT_FOUND,
-  USER_FIELDS,
-  REFRESH_TOKEN_EXPIRED,
-} from './constants';
 import { hashPassword, comparePasswords } from '../../../utils/salt';
 
 export default async function router(instance) {
   const {
+    config: {
+      userService: {
+        userServiceErrors: errors,
+        userServiceMessages: messages,
+        userServiceConstants: constants,
+      },
+      globals: { roles, searchLimits, patterns },
+    },
     models: { User, UserRole },
   } = instance;
 
@@ -43,13 +27,24 @@ export default async function router(instance) {
     method: 'POST',
     url: '/signup',
     schema: {
-      body: signupBodyValidator,
+      body: {
+        type: 'object',
+        properties: {
+          email: { type: 'string' },
+          password: { type: 'string', pattern: patterns.PASSWORD },
+          firstName: { type: 'string' },
+          lastName: { type: 'string' },
+        },
+        required: ['email', 'password', 'firstName', 'lastName'],
+      },
       response: {
         '4xx': { $ref: '4xx#' },
         '5xx': { $ref: '5xx#' },
       },
     },
-    validatorCompiler,
+    preHandler: async ({ body: { email } }) => {
+      instance.validateEmail({ email });
+    },
     handler: async (req, repl) => {
       req.body.password = await hashPassword(req.body.password);
 
@@ -64,7 +59,7 @@ export default async function router(instance) {
           refreshToken,
         });
       } catch (err) {
-        throw new UniqueViolationError(USER_ALREADY_REGISTERED);
+        throw new ConflictError(errors.USER_ERR_ALREADY_REGISTERED);
       }
     },
   });
@@ -73,13 +68,19 @@ export default async function router(instance) {
     method: 'POST',
     url: '/signin',
     schema: {
-      body: signinBodyValidator,
+      body: {
+        type: 'object',
+        properties: {
+          email: { type: 'string' },
+          password: { type: 'string' },
+        },
+        required: ['email', 'password'],
+      },
       response: {
         '4xx': { $ref: '4xx#' },
         '5xx': { $ref: '5xx#' },
       },
     },
-    validatorCompiler,
     handler: async (req, repl) => {
       const { email, password } = req.body;
 
@@ -87,12 +88,12 @@ export default async function router(instance) {
         email,
       });
       if (!userData) {
-        throw new AuthorizationError(UNAUTHORIZED);
+        throw new AuthorizationError(errors.USER_ERR_UNAUTHORIZED);
       }
 
       const compareResult = await comparePasswords(password, userData.password);
       if (!compareResult) {
-        throw new AuthorizationError(UNAUTHORIZED);
+        throw new AuthorizationError(errors.USER_ERR_UNAUTHORIZED);
       }
 
       const accessToken = createAccessToken(instance, userData);
@@ -109,26 +110,31 @@ export default async function router(instance) {
     method: 'POST',
     url: '/refresh_token',
     schema: {
-      body: refreshBodyValidator,
+      body: {
+        type: 'object',
+        properties: {
+          refreshToken: { type: 'string' },
+        },
+        required: ['refreshToken'],
+      },
       response: {
         '4xx': { $ref: '4xx#' },
         '5xx': { $ref: '5xx#' },
       },
     },
-    validatorCompiler,
     handler: async (req, repl) => {
       const { refreshToken } = req.body;
 
       const decoded = await instance.jwt.decode(refreshToken);
 
       if (Date.now() >= decoded.exp * 1000) {
-        throw new AuthorizationError(REFRESH_TOKEN_EXPIRED);
+        throw new AuthorizationError(errors.USER_ERR_TOKEN_EXPIRED);
       }
 
       const userData = await User.query().findById(decoded.id);
 
       if (!userData) {
-        throw new AuthorizationError(UNAUTHORIZED);
+        throw new AuthorizationError(errors.USER_ERR_UNAUTHORIZED);
       }
 
       const newAccessToken = createAccessToken(instance, userData);
@@ -150,17 +156,16 @@ export default async function router(instance) {
         '5xx': { $ref: '5xx#' },
       },
     },
-    validatorCompiler,
     async onRequest(req) {
       await instance.auth({ req });
     },
     handler: async (req, repl) => {
       const userData = await User.query()
         .findById(req.user.id)
-        .select([...USER_FIELDS, 'isSuperAdmin']);
+        .select([...constants.USER_CONST_ALLOWED_USER_FIELDS, 'isSuperAdmin']);
 
       if (!userData) {
-        throw new AuthorizationError(UNAUTHORIZED);
+        throw new AuthorizationError(errors.USER_ERR_UNAUTHORIZED);
       }
 
       const { isSuperAdmin, ...user } = userData;
@@ -169,13 +174,13 @@ export default async function router(instance) {
         .for(UserRole.query().where('user_id', req.user.id))
         .select('name');
 
-      const roles = rolesData.map((role) => role.name);
+      const userRoles = rolesData.map((role) => role.name);
 
       if (isSuperAdmin) {
-        roles.push('SuperAdmin');
+        userRoles.push('SuperAdmin');
       }
 
-      const userWithRoles = { ...user, roles };
+      const userWithRoles = { ...user, roles: userRoles };
 
       return repl.status(200).send(userWithRoles);
     },
@@ -190,7 +195,6 @@ export default async function router(instance) {
         '5xx': { $ref: '5xx#' },
       },
     },
-    validatorCompiler,
     async onRequest(req) {
       await instance.auth({ req, isAdminOnly: true });
     },
@@ -215,9 +219,9 @@ export default async function router(instance) {
       const data = await User.query()
         .skipUndefined()
         .select(
-          USER_ADMIN_FIELDS,
+          constants.USER_CONST_ALLOWED_ADMIN_FIELDS,
           User.relatedQuery('users_roles')
-            .where('role_id', config.roles.TEACHER.id)
+            .where('role_id', roles.TEACHER.id)
             .select(
               raw(`CAST(CASE WHEN COUNT(*) > 0 THEN 1 ELSE 0 END AS INT)`),
             )
@@ -251,7 +255,7 @@ export default async function router(instance) {
           id: req.user.id,
         })
         .offset(req.query.offset || 0)
-        .limit(req.query.limit || config.search.USER_SEARCH_LIMIT);
+        .limit(req.query.limit || searchLimits.USER_SEARCH_LIMIT);
 
       const count = await User.query()
         .skipUndefined()
@@ -282,86 +286,130 @@ export default async function router(instance) {
 
   instance.route({
     method: 'GET',
-    url: '/:id',
+    url: '/:userId',
     schema: {
+      params: {
+        type: 'object',
+        properties: {
+          userId: { type: 'number' },
+        },
+        required: ['userId'],
+      },
       response: {
         '4xx': { $ref: '4xx#' },
         '5xx': { $ref: '5xx#' },
       },
     },
-    validatorCompiler,
     async onRequest(req) {
       await instance.auth({ req, isAdminOnly: true });
     },
-    handler: async (req, repl) => {
-      const id = validateId(req.params.id, req.user.id);
-
-      const data = await User.query().findById(id).select(USER_ADMIN_FIELDS);
-      if (!data) {
-        throw new NotFoundError(USER_NOT_FOUND);
+    handler: async ({ params: { userId }, user: { id } }) => {
+      if (userId === id) {
+        throw new BadRequestError(errors.USER_ERR_INVALID_USER_ID);
       }
 
-      return repl.status(200).send({ data });
+      const data = await User.query()
+        .findById(userId)
+        .select(constants.USER_CONST_ALLOWED_ADMIN_FIELDS);
+      if (!data) {
+        throw new NotFoundError(errors.USER_ERR_USER_NOT_FOUND);
+      }
+
+      return { data };
     },
   });
 
   instance.route({
     method: 'PATCH',
-    url: '/:id',
+    url: '/:userId',
     schema: {
-      body: patchBodyValidator,
+      body: {
+        type: 'object',
+        properties: {
+          email: { type: 'string' },
+          password: { type: 'string', pattern: patterns.PASSWORD },
+          firstName: { type: 'string' },
+          lastName: { type: 'string' },
+          isConfirmed: { type: 'boolean' },
+          isSuperAdmin: { type: 'boolean' },
+        },
+      },
+      params: {
+        type: 'object',
+        properties: {
+          userId: { type: 'number' },
+        },
+        required: ['userId'],
+      },
       response: {
         '4xx': { $ref: '4xx#' },
         '5xx': { $ref: '5xx#' },
       },
     },
-    validatorCompiler,
     async onRequest(req) {
       await instance.auth({ req, isAdminOnly: true });
     },
-    handler: async (req, repl) => {
-      const id = validateId(req.params.id, req.user.id);
+    preHandler: async ({ body: { email } }) => {
+      instance.validateEmail({ email });
+    },
+    handler: async ({ params: { userId }, user: { id }, body }) => {
+      if (userId === id) {
+        throw new BadRequestError(errors.USER_ERR_INVALID_USER_ID);
+      }
 
-      if (req.body.password) {
-        req.body.password = await hashPassword(req.body.password);
+      let hash;
+      if (body.password) {
+        hash = await hashPassword(body.password);
       }
 
       const data = await User.query()
-        .patch(req.body)
-        .findById(id)
-        .returning(USER_ADMIN_FIELDS);
+        .skipUndefined()
+        .patch({
+          ...body,
+          password: hash,
+        })
+        .findById(userId)
+        .returning(constants.USER_CONST_ALLOWED_ADMIN_FIELDS);
 
       if (!data) {
-        throw new BadRequestError(INVALID_PATCH);
+        throw new BadRequestError(errors.USER_ERR_INVALID_UPDATE);
       }
 
-      return repl.status(200).send({ data });
+      return { data };
     },
   });
 
   instance.route({
     method: 'DELETE',
-    url: '/:id',
+    url: '/:userId',
     schema: {
+      params: {
+        type: 'object',
+        properties: {
+          userId: { type: 'number' },
+        },
+        required: ['userId'],
+      },
       response: {
         '4xx': { $ref: '4xx#' },
         '5xx': { $ref: '5xx#' },
       },
     },
-    validatorCompiler,
     async onRequest(req) {
       await instance.auth({ req, isAdminOnly: true });
     },
-    handler: async (req, repl) => {
-      const id = validateId(req.params.id, req.user.id);
-
-      const result = await User.query().deleteById(id);
-
-      if (!result) {
-        throw new NotFoundError(USER_NOT_FOUND);
+    handler: async ({ params: { userId }, user: { id } }) => {
+      if (userId === id) {
+        throw new BadRequestError(errors.USER_ERR_INVALID_USER_ID);
       }
 
-      return repl.status(200).send(USER_DELETED);
+      const result = await User.query().deleteById(userId);
+
+      if (!result) {
+        throw new NotFoundError(errors.USER_ERR_USER_NOT_FOUND);
+      }
+
+      return { message: messages.USER_MSG_USER_DELETED };
     },
   });
 
@@ -369,36 +417,43 @@ export default async function router(instance) {
     method: 'POST',
     url: '/appoint_teacher',
     schema: {
-      body: roleBodyValidator,
+      body: {
+        type: 'object',
+        properties: {
+          id: { type: 'number' },
+        },
+        required: ['id'],
+      },
       response: {
         '4xx': { $ref: '4xx#' },
         '5xx': { $ref: '5xx#' },
       },
     },
-    validatorCompiler,
     async onRequest(req) {
       await instance.auth({ req, isAdminOnly: true });
     },
-    handler: async (req, repl) => {
-      const id = validateId(req.body.id, req.user.id);
+    handler: async ({ body: { id }, user: { id: userId } }) => {
+      if (userId === id) {
+        throw new BadRequestError(errors.USER_ERR_INVALID_USER_ID);
+      }
 
       const check = await UserRole.query().findOne({
         userId: id,
-        roleId: config.roles.TEACHER.id,
+        roleId: roles.TEACHER.id,
       });
 
       if (check) {
-        throw new BadRequestError(ALTER_ROLE_FAIL);
+        throw new BadRequestError(errors.USER_ERR_FAIL_ALTER_ROLE);
       }
 
       await UserRole.query()
         .insert({
           userId: id,
-          roleId: config.roles.TEACHER.id,
+          roleId: roles.TEACHER.id,
         })
         .returning('*');
 
-      return repl.status(200).send(ALTER_ROLE_SUCCESS);
+      return { message: messages.USER_MSG_SUCCESS_ALTER_ROLE };
     },
   });
 
@@ -406,29 +461,36 @@ export default async function router(instance) {
     method: 'POST',
     url: '/remove_teacher',
     schema: {
-      body: roleBodyValidator,
+      body: {
+        type: 'object',
+        properties: {
+          id: { type: 'number' },
+        },
+        required: ['id'],
+      },
       response: {
         '4xx': { $ref: '4xx#' },
         '5xx': { $ref: '5xx#' },
       },
     },
-    validatorCompiler,
     async onRequest(req) {
       await instance.auth({ req, isAdminOnly: true });
     },
-    handler: async (req, repl) => {
-      const id = validateId(req.body.id, req.user.id);
+    handler: async ({ body: { id }, user: { id: userId } }) => {
+      if (userId === id) {
+        throw new BadRequestError(errors.USER_ERR_INVALID_USER_ID);
+      }
 
       const result = await UserRole.query().delete().where({
         userId: id,
-        roleId: config.roles.TEACHER.id,
+        roleId: roles.TEACHER.id,
       });
 
       if (!result) {
-        throw new NotFoundError(USER_ROLE_NOT_FOUND);
+        throw new NotFoundError(errors.USER_ERR_ROLE_NOT_FOUND);
       }
 
-      return repl.status(200).send(ALTER_ROLE_SUCCESS);
+      return { message: messages.USER_MSG_SUCCESS_ALTER_ROLE };
     },
   });
 }
